requirements.txt
alembic==1.10.4
# authlib==1.3.1
fastapi==0.95.1
fastapi_jwt==0.3.0
uvicorn==0.22.0
sqlalchemy==1.4.46
python-jose==3.3.0
psycopg2-binary==2.9.6
pydantic==1.10.8
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
celery==5.2.7
python-dotenv==1.0.1


